+++ lib/kronos/logger.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Kronos
+  module Logger
+  end
+end
+
+require 'kronos/logger/stdout'